vmm_tests: vmbus relay test for TDX (#1472) (#1672)
Fix for issue #1268

- Added changes to petri to be able to run test vm with vmbus relay
enabled.
- Implement the petri vmbus relay config for Hyper-V and OpenVMM
- Added a new boot test with vmbus relay enable for TDX.

Cherry-pick of #1472

---------
