+++ petri/Cargo.toml
@@ -57,6 +57,7 @@ mesh_worker.workspace = true
 mesh.workspace = true
 pal_async.workspace = true
 pal.workspace = true
+powershell_builder.workspace = true
 unix_socket.workspace = true
 sparse_mmap.workspace = true
