+++ petri/src/vm/openvmm/mod.rs
@@ -190,6 +190,14 @@ impl PetriVmConfig for PetriVmConfigOpenVmm {
     fn with_uefi_frontpage(self: Box<Self>, enable: bool) -> Box<dyn PetriVmConfig> {
         Box::new(Self::with_uefi_frontpage(*self, enable))
     }
+
+    fn with_vmbus_redirect(self: Box<Self>, _: bool) -> Box<dyn PetriVmConfig> {
+        Box::new(Self::with_vmbus_redirect(*self))
+    }
+
+    fn os_flavor(&self) -> OsFlavor {
+        self.firmware.os_flavor()
+    }
 }
 
 /// Various channels and resources used to interact with the VM while it is running.
