+++ support/powershell_builder/Cargo.toml
@@ -1,0 +1,14 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+[package]
+name = "powershell_builder"
+edition.workspace = true
+rust-version.workspace = true
+
+[target.'cfg(windows)'.dependencies]
+guid.workspace = true
+jiff.workspace = true
+
+[lints]
+workspace = true
